fixed token validation
