+++ app/_api/api.js
@@ -1,6 +1,5 @@
 import axios from 'axios';
-import cookie from 'js-cookie';
-import ServerCookie from 'next-cookies';
+import { getJwtToken } from '../functions/tokenFunctions';
 
 export const getBaseUrl = () => {
     return process.browser ?
@@ -8,46 +7,10 @@ export const getBaseUrl = () => {
         process.env.BASE_URL || 'http://cms:1337';
 }
 
-
-function parseJwt(token) {
-    if (!token) return;
-    var base64Url = token.split('.')[1];
-    var base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-    var jsonPayload = decodeURIComponent(atob(base64).split('').map(function (c) {
-        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-    }).join(''));
-
-    return JSON.parse(jsonPayload);
-};
-const validateCookie = (token) => {
-    try {
-        const parsed = parseJwt(token);
-        return true;
-    } catch (e) {
-        console.error('error in cookie', e);
-        return false;
-    }
-}
-const getCookie = (ctx) => {
-    if (ctx) {
-        return ServerCookie(ctx).jwt;
-    } else {
-        const jwt = cookie.get('jwt');
-        if (validateCookie(jwt)) {
-            return jwt;
-        }
-
-        if (jwt && !ctx) {
-            cookie.remove('jwt');
-        }
-    }
-    return false;
-}
-
 export const getApi = (ctx) => axios.create({
     baseURL: getBaseUrl(),
     transformRequest: [function (data, headers) {
-        const jwt = getCookie(ctx);
+        const jwt = getJwtToken(ctx);
         if (jwt) {
             headers.common['Authorization'] = `Bearer ${jwt}`;
             //headers.common['cookie'] =
