+++ app/hooks/useSelf.js
@@ -5,8 +5,8 @@ import useSWR from 'swr';
 export function useSelf() {
     const api = GetUsersApi();
     const self = useSWR('/me', path => {
-        return api.self().catch(() => {
-            console.error('error');
+        return api.self().catch((error) => {
+            console.error('error', error);
         })
     });
     return [self.data, self.revalidate];
