+++ app/pages/login.js
@@ -1,11 +1,11 @@
-import nextCookie from 'next-cookies';
 import { Container, Image, Segment, Transition } from 'semantic-ui-react';
-import { GetPageSettingsApi } from '../_api';
+import { ForgotPassword } from '../components/auth/ForgotPassword';
 import { LoginForm } from '../components/auth/LoginForm';
 import { ResetPassword } from '../components/auth/ResetPassword';
 import { absUrl } from '../functions/imageFunctions';
+import { getJwtToken, validateToken } from '../functions/tokenFunctions';
 import { useLogin } from '../hooks/useLogin';
-import { ForgotPassword } from '../components/auth/ForgotPassword';
+import { GetPageSettingsApi } from '../_api';
 
 const Page = function ({ bannerImage }) {
     const [state] = useLogin();
@@ -38,14 +38,15 @@ const Page = function ({ bannerImage }) {
 const LoginPage = Page;
 
 LoginPage.getInitialProps = async ctx => {
-    const cookies = nextCookie(ctx);
-    const jwt = cookies.jwt;
-    const { bannerImage } = await GetPageSettingsApi(ctx).get('Login');
-    if (ctx.req && jwt) {
+    const jwt = getJwtToken(ctx);
+
+    if (validateToken(jwt)) {
         ctx.res.writeHead(302, { Location: '/' })
         ctx.res.end()
-        return
+        return;
     }
+
+    const { bannerImage } = await GetPageSettingsApi(ctx).get('Login');
     return {
         jwt,
         bannerImage
